requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
langchain>=0.1.0
langchain-openai>=0.0.5
chromadb>=0.4.22
pypdf>=4.0.0
python-dotenv>=1.0.0